refactor: merge src/utils into src/common/utils
