+++ src/common/toast.tsx
@@ -1,7 +1,7 @@
 import React, { ReactNode } from 'react'
 import toast from 'react-hot-toast'
 import { playAlarm } from './playAlarm'
-import { translateError } from '@/utils/translate-error'
+import { translateError } from '@/common/utils/translate-error'
 import { Icon } from '../icons'
 
 type ToastType = 'success' | 'error' | 'info' | 'warning'
