+++ src/context/wallpaper.context.tsx
@@ -5,7 +5,7 @@ import { callEvent, listenEvent } from '@/common/utils/call-event'
 import type { StoredWallpaper, Wallpaper } from '@/common/wallpaper.interface'
 import { safeAwait } from '@/services/api'
 import { useChangeWallpaper } from '@/services/hooks/extension/updateSetting.hook'
-import { translateError } from '@/utils/translate-error'
+import { translateError } from '@/common/utils/translate-error'
 import Analytics from '@/analytics'
 import { showToast } from '@/common/toast'
 import { useQueryClient } from '@tanstack/react-query'
