+++ src/layouts/friends/components/add-friend.bottomSheet.tsx
@@ -3,7 +3,7 @@ import { Button } from '@/components/button/button'
 import { TextInput } from '@/components/text-input'
 import { useAuth } from '@/context/auth.context'
 import { useSendFriendRequest } from '@/services/hooks/friends/friendService.hook'
-import { translateError } from '@/utils/translate-error'
+import { translateError } from '@/common/utils/translate-error'
 import { showToast } from '@/common/toast'
 import Modal from '@/components/modal'
 import { Icon } from '@/src/icons'
