+++ src/layouts/market/other-items/components/market-item-purchase-modal.tsx
@@ -4,7 +4,7 @@ import { ItemPrice } from '@/components/item-price/item-price'
 import Modal from '@/components/modal'
 import type { MarketItem } from '@/services/hooks/market/market.interface'
 import { usePurchaseMarketItem } from '@/services/hooks/market/purchaseMarketItem.hook'
-import { translateError } from '@/utils/translate-error'
+import { translateError } from '@/common/utils/translate-error'
 import { showToast } from '@/common/toast'
 import { RenderPreview } from './renderPreview'
 import { Icon } from '@/src/icons'
