+++ src/layouts/setting/tabs/account/auth-form/components/login-google.button.tsx
@@ -8,7 +8,7 @@ import { useState } from 'react'
 import { safeAwait } from '@/services/api'
 import type { AxiosError } from 'axios'
 import { showToast } from '@/common/toast'
-import { translateError } from '@/utils/translate-error'
+import { translateError } from '@/common/utils/translate-error'
 import Analytics from '@/analytics'
 import { callEvent } from '@/common/utils/call-event'
 import { sleep } from '@/common/utils/timeout'
