+++ src/layouts/setting/tabs/account/components/modals/add-phone.modal.tsx
@@ -1,6 +1,6 @@
 import Modal from '@/components/modal'
 import { TextInput } from '@/components/text-input'
-import { isEmpty, isLessThan } from '@/utils/validators'
+import { isEmpty, isLessThan } from '@/common/utils/validators'
 import InputTextError from '../../auth-form/components/input-text-error'
 import { Button } from '@/components/button/button'
 import OtpInput from '../../auth-form/components/otp-input'
@@ -9,7 +9,7 @@ import {
 	useChangePhoneVerify,
 } from '@/services/hooks/user/userService.hook'
 import { safeAwait } from '@/services/api'
-import { translateError } from '@/utils/translate-error'
+import { translateError } from '@/common/utils/translate-error'
 import { showToast } from '@/common/toast'
 
 interface AddPhoneProp {
