+++ src/layouts/setting/tabs/account/components/modals/edit-username.tsx
@@ -5,7 +5,7 @@ import { TextInput } from '@/components/text-input'
 import { safeAwait } from '@/services/api'
 import { useUpdateUsername } from '@/services/hooks/auth/authService.hook'
 import type { UserProfile } from '@/services/hooks/user/userService.hook'
-import { translateError } from '@/utils/translate-error'
+import { translateError } from '@/common/utils/translate-error'
 import type { AxiosError } from 'axios'
 import { useState } from 'react'
 import { FooterButtons } from './footer-buttons'
