+++ src/layouts/widgets/habit/components/item/habit.item.tsx
@@ -7,7 +7,7 @@ import type { WidgetifyDate } from '@/layouts/widgets/calendar/utils'
 import { safeAwait } from '@/services/api'
 import { HabitComparison, type Habit } from '@/services/hooks/habit/habit.interface'
 import { useLogHabitProgress } from '@/services/hooks/habit/log-habit-progress.hook'
-import { translateError } from '@/utils/translate-error'
+import { translateError } from '@/common/utils/translate-error'
 import { formatHabitGoal } from '../../utils'
 import { SegmentedProgressRing } from './button.progress-ring'
 import { SimpleProgressRing } from './button.simple-progress-ring'
