+++ src/layouts/widgets/todos/expandable-todo-input.tsx
@@ -16,7 +16,7 @@ import { PriorityDropdown } from './components/priority.dropdown'
 import type { FetchedTodo, TodoPriority } from '@/services/hooks/todo/todo.interface'
 import { type TodoCreationPayload, useAddTodo } from '@/services/hooks/todo/add-todo.hook'
 import { useUpdateTodo } from '@/services/hooks/todo/update-todo.hook'
-import { translateError } from '@/utils/translate-error'
+import { translateError } from '@/common/utils/translate-error'
 import { showToast } from '@/common/toast'
 import type { Friend } from '@/services/hooks/friends/friendService.hook'
 import { TodoSelectFriends } from './components/select-friends.todo'
